workshops/MCP_workshop/pyproject.toml
[project]
name = "mcp-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.0.2",
]
